requirements.txt
ccxt>=4.0.0
anthropic>=0.40.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
websockets>=12.0
pydantic>=2.0.0
loguru>=0.7.0
